+++ packages/components/avatar/src/use-avatar.ts
@@ -4,9 +4,8 @@ import type {ReactRef} from "@heroui/react-utils";
 
 import {avatar} from "@heroui/theme";
 import {useProviderContext} from "@heroui/system";
-import {mergeProps} from "@react-aria/utils";
 import {useDOMRef, filterDOMProps} from "@heroui/react-utils";
-import {clsx, safeText, dataAttr} from "@heroui/shared-utils";
+import {clsx, safeText, dataAttr, mergeProps} from "@heroui/shared-utils";
 import {useFocusRing} from "@react-aria/focus";
 import {useMemo, useCallback} from "react";
 import {useImage} from "@heroui/use-image";